requirements.txt
torch
torchaudio
deepfilternet
ffmpeg-python